requirements.txt
asgiref==3.2.10
cycler==0.10.0
Django==3.0.8
future==0.18.2
image==1.5.32
joblib==0.16.0
kiwisolver==1.2.0
matlab==0.1
matplotlib==3.3.0
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
PyQt5-stubs==5.14.2.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
scikit-learn==0.23.1
scipy==1.5.2
seaborn==0.10.1
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.6.0+cu101
torchvision==0.7.0+cu101
